+++ RELEASES.md
@@ -43,6 +43,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates: refactored to rely on service-templates-core
 * prov-template: template variable syntax checker
 * prov-template-compiler: introduced fully qualified names for templates
+* prov-template-compiler: LocalEnactor for composites
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java